Added name and some taboo, added some tests
Solver: Added my name and have a taboo cells function that can satisfy
rule 1 but not 2 yet (work in progress). It may not be the most elegant
solution yet but I try

Tester: Added an extra test for the taboo cells for Rule 2. Finds the
corners but still doesn't work though